update reservations controller

+++ app/controllers/reservations_controller.rb
@@ -1,5 +1,5 @@
 class ReservationsController < ApplicationController
-  before_action :authorize, only: %i[create destroy]
+  before_action :authorize, only: %i[create update destroy]
   before_action :read_reservation, only: [:destroy]
 
   def index
@@ -10,22 +10,34 @@ class ReservationsController < ApplicationController
   def create
     @reservation = Reservation.new(reservation_params)
     @reservation.user = @user
+    car = Car.where(id: params[:reservation][:car_id])
     if @reservation.save
-      render json: @reservation
+      render json: { reservation: @reservation, name: @user.name,
+                     car: }
     else
       render json: { error: @reservation.errors.full_messages }
     end
   end
 
   def update
+    unless @user == Reservation.find(params[:id]).user
+      return render json: { error: 'You are not allowed because you did not create the reservation' },
+                    status: :unauthorized
+    end
+    car = Car.where(id: params[:reservation][:car_id])
     if @reservation.update(reservation_params)
-      render json: @reservation
+      render json: { reservation: @reservation, name: @user.name,
+                     car: }
     else
       render json: @reservation.errors.full_messages
     end
   end
 
   def destroy
+    unless @user == Reservation.find(params[:id]).user
+      return render json: { error: 'You are not allowed because you did not create the reservation' },
+                    status: :unauthorized
+    end
     if @reservation.destroy
       render json: { id: @reservation.id, message: 'Reservation deleted successfully' }
     else
